Add LOAD OPETREE and LOAD SCHEMA

+++ PrismaDB-QueryParser-MSSQL/Helpers/CaseChangingCharStream.cs
@@ -68,10 +68,10 @@ namespace PrismaDB.QueryParser.MSSQL
 
             if (upper)
             {
-                return (int)char.ToUpperInvariant(o);
+                return char.ToUpperInvariant(o);
             }
 
-            return (int)char.ToLowerInvariant(o);
+            return char.ToLowerInvariant(o);
         }
 
         public int Mark()
